Object.assing polyfill moved to utils
