+++ modules/core/index.ts
@@ -4,34 +4,10 @@ export * from "./src/dynamic-input-control.model";
 export * from "./src/dynamic-option-control.model";
 export * from "./src/dynamic-form.service";
 export * from "./src/dynamic-form-control.component";
+export * from "./src/utils";
 
 export * from "./src/checkbox/dynamic-checkbox.model";
 export * from "./src/input/dynamic-input.model";
 export * from "./src/radio/dynamic-radio.model";
 export * from "./src/select/dynamic-select.model";
 export * from "./src/textarea/dynamic-textarea.model";
-
-// Object.assign Polyfill 
-if (typeof Object.assign != 'function') {
-    (function () {
-        Object.assign = function (target) {
-            'use strict';
-            if (target === undefined || target === null) {
-                throw new TypeError('Cannot convert undefined or null to object');
-            }
-
-            var output = Object(target);
-            for (var index = 1; index < arguments.length; index++) {
-                var source = arguments[index];
-                if (source !== undefined && source !== null) {
-                    for (var nextKey in source) {
-                        if (source.hasOwnProperty(nextKey)) {
-                            output[nextKey] = source[nextKey];
-                        }
-                    }
-                }
-            }
-            return output;
-        };
-    })();
-}
